+++ packages/medusa/src/api/store/customers/validators.ts
@@ -10,7 +10,7 @@ export const StoreCreateCustomer = z.object({
   first_name: z.string().nullish(),
   last_name: z.string().nullish(),
   phone: z.string().nullish(),
-  metadata: z.record(z.unknown()).nullish(),
+  metadata: z.record(z.string(), z.unknown()).nullish(),
 })
 
 export const StoreUpdateCustomer = z.object({
@@ -18,7 +18,7 @@ export const StoreUpdateCustomer = z.object({
   first_name: z.string().nullish(),
   last_name: z.string().nullish(),
   phone: z.string().nullish(),
-  metadata: z.record(z.unknown()).nullish(),
+  metadata: z.record(z.string(), z.unknown()).nullish(),
 })
 
 export const StoreGetCustomerAddressParams = createSelectParams()