+++ packages/modules/link-modules/src/definitions/user-rbac-role.ts
@@ -1,5 +1,10 @@
 import { ModuleJoinerConfig } from "@medusajs/framework/types"
-import { LINKS, Modules } from "@medusajs/framework/utils"
+import {
+  defineFileConfig,
+  FeatureFlag,
+  LINKS,
+  Modules,
+} from "@medusajs/framework/utils"
 
 export const UserRbacRole: ModuleJoinerConfig = {
   serviceName: LINKS.UserRbacRole,
@@ -72,3 +77,7 @@ export const UserRbacRole: ModuleJoinerConfig = {
     },
   ],
 }
+
+defineFileConfig({
+  isDisabled: () => !FeatureFlag.isFeatureEnabled("rbac"),
+})