+++ www/apps/bloom/sidebars/features.mjs
@@ -19,11 +19,38 @@ export const featuresSidebar = [
             title: "View Switcher",
             path: "/features/view-switcher",
           },
-          // {
-          //   type: "link",
-          //   title: "Emails",
-          //   path: "/features/emails",
-          // },
+          {
+            type: "link",
+            title: "Emails",
+            path: "/features/emails",
+          },
+          {
+            type: "link",
+            title: "Integrations",
+            path: "/features/integrations",
+            children: [
+              {
+                type: "link",
+                title: "Cookiebot",
+                path: "/features/integrations/guides/cookiebot",
+              },
+              {
+                type: "link",
+                title: "Klaviyo",
+                path: "/features/integrations/guides/klaviyo",
+              },
+              {
+                type: "link",
+                title: "Slack",
+                path: "/features/integrations/guides/slack",
+              },
+              {
+                type: "link",
+                title: "Stripe",
+                path: "/features/integrations/guides/stripe",
+              },
+            ],
+          },
           {
             type: "link",
             title: "Selection Mode",